feat(#120 server/realtime): move Field to common
