+++ modules/matches/Realtime/client/rust/Realtime/src/ffi/command/event.rs
@@ -1,6 +1,6 @@
 use cheetah_matches_realtime_common::commands::c2s::C2SCommand;
+use cheetah_matches_realtime_common::commands::field::FieldId;
 use cheetah_matches_realtime_common::commands::types::event::{EventCommand, TargetEventCommand};
-use cheetah_matches_realtime_common::constants::FieldId;
 use cheetah_matches_realtime_common::room::RoomMemberId;
 
 use crate::clients::registry::ClientId;
