+++ modules/matches/Realtime/client/rust/Realtime/tests/emulation.rs
@@ -6,7 +6,7 @@ use lazy_static::lazy_static;
 use cheetah_matches_realtime_client::ffi;
 use cheetah_matches_realtime_client::ffi::channel::Channel;
 use cheetah_matches_realtime_client::ffi::GameObjectIdFFI;
-use cheetah_matches_realtime_common::constants::FieldId;
+use cheetah_matches_realtime_common::commands::field::FieldId;
 use cheetah_matches_realtime_common::room::RoomMemberId;
 
 use crate::helpers::helper::setup;
