+++ modules/matches/Realtime/common/Common/src/commands/mod.rs
@@ -4,12 +4,12 @@ use crate::protocol::codec::commands::context::CommandContextError;
 
 pub mod binary_value;
 pub mod c2s;
-pub mod field_type;
+pub mod field;
 mod field_value;
 pub mod s2c;
 pub mod types;
 
-pub use crate::commands::field_type::FieldType;
+pub use crate::commands::field::FieldType;
 pub use crate::commands::field_value::FieldValue;
 
 ///
