+++ modules/matches/Realtime/server/src/room/command/field.rs
@@ -1,9 +1,10 @@
+use cheetah_matches_realtime_common::commands::field::Field;
 use cheetah_matches_realtime_common::commands::s2c::S2CCommand;
 use cheetah_matches_realtime_common::commands::types::field::DeleteFieldCommand;
 use cheetah_matches_realtime_common::room::RoomMemberId;
 
 use crate::room::command::{ServerCommandError, ServerCommandExecutor};
-use crate::room::object::{Field, GameObject};
+use crate::room::object::GameObject;
 use crate::room::template::config::Permission;
 use crate::room::Room;
