+++ modules/matches/Realtime/server/src/room/command/long.rs
@@ -1,3 +1,4 @@
+use cheetah_matches_realtime_common::commands::field::Field;
 use cheetah_matches_realtime_common::commands::s2c::S2CCommand;
 use cheetah_matches_realtime_common::commands::types::field::SetFieldCommand;
 use cheetah_matches_realtime_common::commands::types::long::IncrementLongC2SCommand;
@@ -5,7 +6,7 @@ use cheetah_matches_realtime_common::commands::FieldType;
 use cheetah_matches_realtime_common::room::RoomMemberId;
 
 use crate::room::command::{ServerCommandError, ServerCommandExecutor};
-use crate::room::object::{Field, GameObject};
+use crate::room::object::GameObject;
 use crate::room::template::config::Permission;
 use crate::room::Room;
